added FileLoggerTests

+++ Logger/FileLogger.cs
@@ -13,21 +13,26 @@ namespace Logger
             {
                 return _filePath!;
             }
-            set => _filePath = value ?? throw new ArgumentNullException();
+            set => _filePath = value??throw new ArgumentNullException();
         }
 
 
         public FileLogger(string filePath)
         {
 
-            this._filePath = filePath;
+            FilePath = filePath;
         }
 
         override
         public void Log(LogLevel logLevel, string message)
         {
+            if(!File.Exists(FilePath))
+            {
+                throw new FileNotFoundException();
+            }
+
             TextWriter tw = new StreamWriter(FilePath, true);
-            
+
             string res = String.Format("{0} {1} {2}: {3}", DateTime.Now.ToString(), base.ClassName, logLevel, message);
 
             tw.WriteLine(res);
